Very basic struture and build script configured.
