+++ .gitignore
@@ -11,6 +11,8 @@ pids
 *.seed
 *.pid.lock
 
+lib
+
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
