+++ config/config.dev.json
@@ -1,0 +1,7 @@
+{
+  "courseScrapeMeta": {
+    "page": {
+      "PAGE_SIZE": 10
+    }
+  }
+}
