+++ src/courseScraper/index.js
@@ -1,0 +1,3 @@
+import CourseScraper from './courseScraper';
+
+export default CourseScraper;
